Enable Keycloak authentication (#3)
Enables Keycloak authentication, which is configured to connect to the BTS Keycloak server (keycloak.bts-crew.com) and we will have 2 realms:

* nonprod for local development and the staging site
* prod for the production site

The Keycloak adaptor only provides us with Role-Based Access Control (RBAC), where you can specify which roles a user can have in order to access an endpoint. However, we often want to authorise based on an action on a given entity (eg, whether you're the "author") and so this also adds a Policy helper to allow domains to authorise based on arbitrary rules. This helper uses Higher-Order Functions (HOF) and so is very flexible; see the docs attached for info on how you can use it.